Most downloaded movies of all time
Step 1: Find the URL that you want to scrape
In my code, I used this URL https://yts.lt/browse-movies/0/all/all/0/downloads it is a torrenting site that you can download movies.
Step 2: Inspecting the Page
By inspecting elements you can the codes of the data you want to be extracted, in my case I used SelectorGadget an google extension so I can easily find the element or the code of the data I wanted to be extracted.
Step 3: Find the data you want to extract
In my case, I did find the movie title, movie rating, movie year and movie genre
Step 4: Write the code
When you type scrapy createproject "name" there will be a code for scraping and it will be easy to manipulate it
Step 5: Store the data in a required format
This format varies depending on your requirement. In my case, I used the code scrapy crawl "yourcode" -o "name".csv it can be viewed by excell apps.
Step 6: Finalized the datas you want to be extracted.